Behaviors working in TradingRewards.js
